Move bulk of attemptDispatchHomeScreen to DispatchActivity
